Create Pointer_Easy.cpp
The user provides two values a and b which are input to the code. These values are stored in pointers pa and pb and passed into the update() function as parameters. The update() function accepts two pointers and doesn't return any value. But as the parameters are pointers, the values can directly be changed in memory as a result the printf() function in main will output the updated value in memory.
Objective:
i)a = a + b
ii)b = |a - b|
for ex, if a = 5; b = 4 then
i)9
ii)1