Create Functions.cpp
This .cpp code takes four integers as an input from the user and passes it into the calling function max_of_four(). The max_of_four() returns an integer which is the largest of all 4 values. The function definition uses deque to find the largest value with a time complexity of O(nlogn).